requirements.txt
Django>=1.7,<1.8
django-ckeditor
django-grappelli
django-smuggler
django-recaptcha
Pillow
psycopg2
wsgiref==0.1.2
django-cities
django-compressor
